requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
defusedxml==0.6.0         # via python3-openid
django-allauth==0.42.0    # via -r requirements.in
django-filter==2.3.0      # via -r requirements.in
django-webpack-loader==0.7.0  # via -r requirements.in
django==3.1.1             # via -r requirements.in, django-allauth, django-filter, djangorestframework
djangorestframework==3.11.1  # via -r requirements.in
idna==2.10                # via requests
oauthlib==3.1.0           # via requests-oauthlib
python3-openid==3.2.0     # via django-allauth
pytz==2020.1              # via django
requests-oauthlib==1.3.0  # via django-allauth
requests==2.24.0          # via django-allauth, requests-oauthlib
sqlparse==0.3.1           # via django
urllib3==1.25.10          # via requests
